refactor(cli): move manifest writes into manifest_write
manifest.py crossed the 700-line review threshold after the flow-range
and registration work. Extract the write/mutation/registration half
(build_manifest_document, write_manifest_if_absent, update_manifest,
write_back_workspace_status, manifest_range_for_flow,
pre_register_workspace) into chipcompiler.cli.project.manifest_write,
keeping manifest.py on loading, normalization, and classification, and
move the write-side tests and monkeypatch targets with the code.
